fix: enforce auth on GET/HEAD requests when credentials configured (#164)
Security improvements from PR consolidation:
- platform_auth: Require auth for all methods (except OPTIONS preflight)
  when credentials are actually configured
- security: Add can_enforce_auth() to check mode-appropriate credentials
  - Basic mode only enforces when token is set (non-empty)
  - JWT mode only enforces when jwt_secret is set (non-empty)
- Tests updated to verify new auth behavior for GET requests

Also includes:
- jwt_validation.rs: Fix type mismatch (u64 vs usize)
- ui_blocking_reproduction.rs: Clippy fix for let_unit_value

+++ crates/app-http/tests/jwt_validation.rs
@@ -11,8 +11,8 @@ use std::time::{SystemTime, UNIX_EPOCH};
 #[derive(Serialize)]
 struct TestClaims {
     sub: String,
-    exp: usize,
-    iat: usize,
+    exp: u64,
+    iat: u64,
     iss: String,
 }
 
@@ -27,8 +27,8 @@ fn create_jwt_token(
 
     let claims = TestClaims {
         sub: subject.to_string(),
-        exp: (now + expires_in_seconds) as usize,
-        iat: now as usize,
+        exp: now + expires_in_seconds,
+        iat: now,
         iss: issuer.to_string(),
     };
